DDoS protection now fully functional!
- Added DDoS protection loop
- Fixed debug bugs
