+++ private/controlpanel-server.js
@@ -11,25 +11,6 @@ var exec = require('child_process').exec;
 var values = mods.values;
 var props = mods.props;
 
-// Get line number; for debugging
-Object.defineProperty(global, '__stack', {
-  get: function(){
-    var orig = Error.prepareStackTrace;
-    Error.prepareStackTrace = function(_, stack){ return stack; };
-    var err = new Error;
-    Error.captureStackTrace(err, arguments.callee);
-    var stack = err.stack;
-    Error.prepareStackTrace = orig;
-    return stack;
-  }
-});
-
-Object.defineProperty(global, '__line', {
-  get: function(){
-    return __stack[1].getLineNumber();
-  }
-});
-
 function printError(reason, id) {
 	io.emit('server-checked', {"success": false, "reason": reason, "id": id});
 }
