+++ private/createserv-server.js
@@ -10,25 +10,6 @@ var mkdir = require('mkdirp');
 
 var doneSearching = false;
 
-// Get line number; for debugging
-Object.defineProperty(global, '__stack', {
-  get: function(){
-    var orig = Error.prepareStackTrace;
-    Error.prepareStackTrace = function(_, stack){ return stack; };
-    var err = new Error;
-    Error.captureStackTrace(err, arguments.callee);
-    var stack = err.stack;
-    Error.prepareStackTrace = orig;
-    return stack;
-  }
-});
-
-Object.defineProperty(global, '__line', {
-  get: function(){
-    return __stack[1].getLineNumber();
-  }
-});
-
 function printError(reason, id) {
 	io.emit('creation-complete', {"success": false, "reason": reason, "id": id});
 }
