+++ private/main.js
@@ -1,5 +1,6 @@
 var toobusy = require('toobusy-js');
 var mods = require('./getProps.js');
+var fs = require('fs');
 var app = mods.app;
 
 app.use(function(req, res, next) {
@@ -10,7 +11,49 @@ app.use(function(req, res, next) {
 	}
 });
 
+// Get line number; for debugging
+Object.defineProperty(global, '__stack', {
+  get: function(){
+    var orig = Error.prepareStackTrace;
+    Error.prepareStackTrace = function(_, stack){ return stack; };
+    var err = new Error;
+    Error.captureStackTrace(err, arguments.callee);
+    var stack = err.stack;
+    Error.prepareStackTrace = orig;
+    return stack;
+  }
+});
+
+Object.defineProperty(global, '__line', {
+  get: function(){
+    return __stack[1].getLineNumber();
+  }
+});
+
 var file_reg = require('./register.js');
-var file_CreateServ = require('./create-server.js');
+// var file_createServ = require('./createserv-server.js');
 var file_login = require('./login.js');
-var cp_serv = require('./controlpanel-server.js');
+var file_cpServ = require('./controlpanel-server.js');
+
+function Reset() {
+	fs.readFile('bans.txt', 'utf8', function(err, data) {
+		if(data != '') {
+			var items = data.split("\n");
+			for(i = 0; i < items.length; i++) {
+				items[i] = items[i].split(" ");
+			}
+			
+			for(i = 0; i < items.length; i++) {
+				items[i][1] = 0;
+				items[i] = items[i].join(" ");
+			}
+			
+			data = items.join("\n");
+			fs.writeFile('bans.txt', data, function(err, data) {
+				return;
+			});
+		}
+	});
+}
+
+setInterval(Reset, 1000);
